Changed variable type to long so it handles the arguments max values

+++ bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned int count = 0;
-	unsigned int xor = n ^ m;
+	unsigned long int xor = n ^ m;
 
 	while (xor > 0)
 	{
